Ditch docker.py and random.py
